proper goals

+++ scripts/core/node/intersection.gd
@@ -4,6 +4,10 @@ var _current: int = 0
 
 
 func _receive_signal(signal_info: SignalInfo) -> void:
+	if outgoing_connections.is_empty():
+		Level.current.stop_simulation()
+		return
+
 	SignalSender.send_signal(position, outgoing_connections[_current].target, signal_info)
 	_current = (_current + 1) % outgoing_connections.size()
